+++ src/ResetCtrl.java
@@ -13,11 +13,11 @@ public class ResetCtrl implements Controller {
 		models.add(model);
 	}
 
-	public void updateModels(Integer num) {
+	public void updateModels(Double num) {
 
 		//New ArrayList resets model in Numbers.java
 		for (Model model: models) {
-			model.update(new ArrayList<Integer>());
+			model.update(new ArrayList<Double>());
 		}
 	}
